refactor: Participant 골격 인터페이스에서 추상 클래스로 변경
